pyproject.toml
[tool.poetry]
name = "p"
version = "0.1.0"
description = ""
authors = ["Fedorov <atiliogoulart@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
pre-commit = "^2.15.0"
djangorestframework-api-key = "2.*"
django-filter = "^21.1"
djangorestframework-swagger = "^0.2.2"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.20.0"
pytest-django = "^4.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
